requirements-vertex.txt
google-cloud-aiplatform>=1.25.0
vertexai>=1.43.0
fastapi>=0.68.0
uvicorn>=0.15.0
python-dotenv>=0.19.0
pydantic>=1.8.0
pandas>=1.3.0
numpy>=1.21.0
scikit-learn>=0.24.0
jupyter>=1.0.0
ipykernel>=6.0.0
requests>=2.26.0
